还需要解决粘包拆包问题

+++ src/main/java/com/hongframe/sekiro/CommandCode.java
@@ -6,6 +6,6 @@ package com.hongframe.sekiro;
  */
 public interface CommandCode {
 
-    int code();
+    short code();
 
 }
